+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/SentenceSentiment.cs
@@ -16,11 +16,13 @@ namespace Azure.AI.TextAnalytics
     /// </summary>
     public readonly struct SentenceSentiment
     {
-        internal SentenceSentiment(TextSentiment sentiment, string text, double positiveScore, double neutralScore, double negativeScore, IReadOnlyList<MinedOpinion> minedOpinions)
+        internal SentenceSentiment(TextSentiment sentiment, string text, double positiveScore, double neutralScore, double negativeScore, int offset, int length, IReadOnlyList<MinedOpinion> minedOpinions)
         {
             Sentiment = sentiment;
             Text = text;
             ConfidenceScores = new SentimentConfidenceScores(positiveScore, neutralScore, negativeScore);
+            Offset = offset;
+            Length = length;
             MinedOpinions = new List<MinedOpinion>(minedOpinions);
         }
 
@@ -33,6 +35,8 @@ namespace Azure.AI.TextAnalytics
             ConfidenceScores = sentenceSentiment.ConfidenceScores;
             Sentiment = (TextSentiment)Enum.Parse(typeof(TextSentiment), sentenceSentiment.Sentiment, ignoreCase: true);
             MinedOpinions = ConvertToMinedOpinions(sentenceSentiment, allSentences);
+            Offset = sentenceSentiment.Offset;
+            Length = sentenceSentiment.Length;
         }
 
         /// <summary>
@@ -56,6 +60,16 @@ namespace Azure.AI.TextAnalytics
         /// <see cref="AnalyzeSentimentOptions.IncludeOpinionMining"/> is set to true.
         /// </summary>
         public IReadOnlyCollection<MinedOpinion> MinedOpinions { get; }
+
+        /// <summary>
+        /// Gets the starting position (in UTF16 code units) for the matching text in the sentence.
+        /// </summary>
+        public int Offset { get; }
+
+        /// <summary>
+        /// Gets the length (in UTF16 code units) of the matching text in the sentence.
+        /// </summary>
+        public int Length { get; }
 
         private static IReadOnlyCollection<MinedOpinion> ConvertToMinedOpinions(SentenceSentimentInternal sentence, IReadOnlyList<SentenceSentimentInternal> allSentences)
         {