+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/RecognizeEntitiesTests.cs
@@ -49,6 +49,9 @@ namespace Azure.AI.TextAnalytics.Tests
             {
                 Assert.IsTrue(entitiesList.Contains(entity.Text));
                 Assert.IsNotNull(entity.ConfidenceScore);
+                Assert.IsNotNull(entity.Offset);
+                Assert.IsNotNull(entity.Length);
+                Assert.Greater(entity.Length, 0);
             }
         }